Ollama_langchain/requirements.txt
python-dotenv
streamlit
langchain-community